+++ src/main/java/models/Game.java
@@ -10,11 +10,8 @@ import java.util.Random;
 public class Game {
 
     public boolean error;
-
     public java.util.List<Card> deck = new ArrayList<>();
-
     public java.util.List<java.util.List<Card>> cols = new ArrayList<>();
-
 
     public Game(){
         cols.add(new ArrayList<Card>());
@@ -24,16 +21,9 @@ public class Game {
         error = false;
     }
 
-
-    public void buildDeck() {
-        for(int i = 2; i < 15; i++){
-            deck.add(new Card(i,Suit.Clubs));
-            deck.add(new Card(i,Suit.Hearts));
-            deck.add(new Card(i,Suit.Diamonds));
-            deck.add(new Card(i,Suit.Spades));
-        }
-    }
-
+    //
+    // buildDeck() will need to be implemented seperately on each subclass
+    //
 
     public void shuffle() {
         long seed = System.nanoTime();
@@ -95,7 +85,6 @@ public class Game {
     private Card getTopCard(int columnNumber) {
         return this.cols.get(columnNumber).get(this.cols.get(columnNumber).size()-1);
     }
-
 
     public void move(int colFrom, int colTo) {
         Card cardToMove = getTopCard(colFrom);
